feat：Move `FeedViewAdapter` to separate file
