+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedViewAdapter.swift
@@ -1,0 +1,31 @@
+//
+//  FeedViewAdapter.swift
+//  EssentialFeediOS
+//
+//  Created by nicaho on 2024/10/21.
+//
+
+import UIKit
+import EssentialFeed
+
+final class FeedViewAdapter: FeedView {
+    private weak var controller: FeedViewController?
+    private let imageLoader: FeedImageDataLoader
+    
+    
+    init(controller: FeedViewController, imageLoader: FeedImageDataLoader) {
+        self.controller = controller
+        self.imageLoader = imageLoader
+    }
+    
+    func display(_ viewModel: FeedViewModel) {
+        controller?.tableModel = viewModel.feed.map { model in
+            let adapt = FeedImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: imageLoader)
+            let view = FeedImageCellController(delegate: adapt)
+            
+            adapt.presenter = FeedImagePresenter(view: WeakRefVirtualProxy(view), imageTransformer: UIImage.init)
+            
+            return view
+        }
+    }
+}
